# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API of the Idee project!",
        "title": "Idee API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/drink": {
            "get": {
                "description": "Get available drinks (bars, bistros) in a city",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "graphql"
                ],
                "summary": "Total",
                "parameters": [
                    {
                        "type": "string",
                        "example": "Paris",
                        "name": "city",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/graphql.GetDrinksResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/graphql.Error"
                        }
                    }
                }
            }
        },
        "/enjoy": {
            "get": {
                "description": "Get events|activities from our DATATourisme API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "graphql"
                ],
                "summary": "Enjoy",
                "parameters": [
                    {
                        "type": "string",
                        "name": "city",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/graphql.GetSleepResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/graphql.Error"
                        }
                    }
                }
            }
        },
        "/total": {
            "get": {
                "description": "Get total number of objets in database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "graphql"
                ],
                "summary": "Total",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/graphql.GetTotalResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/graphql.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "graphql.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "string"
                }
            }
        },
        "graphql.GetDrinksPoiPointOfInterest_ResultSet": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphql.GetDrinksPoiPointOfInterest_ResultSetResultsPointOfInterest"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "graphql.GetDrinksPoiPointOfInterest_ResultSetResultsPointOfInterest": {
            "type": "object",
            "properties": {
                "hasContact": {
                    "description": "L\u0026#039;agent à contacter pour affaires générales relatives à ce POI.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphql.GetDrinksPoiPointOfInterest_ResultSetResultsPointOfInterestHasContactAgent"
                    }
                },
                "rdfs_comment": {
                    "description": "Description de la ressource.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphql.GetDrinksPoiPointOfInterest_ResultSetResultsPointOfInterestRdfs_commentLangString"
                    }
                },
                "rdfs_label": {
                    "description": "Etiquette courte décrivant la ressource.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphql.GetDrinksPoiPointOfInterest_ResultSetResultsPointOfInterestRdfs_labelLangString"
                    }
                }
            }
        },
        "graphql.GetDrinksPoiPointOfInterest_ResultSetResultsPointOfInterestHasContactAgent": {
            "type": "object",
            "properties": {
                "foaf_homepage": {
                    "description": "L\u0026#039;adresse du site internet d\u0026#039;un Agent.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "schema_email": {
                    "description": "Un courriel, courrier électronique.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "schema_telephone": {
                    "description": "Un numéro de téléphone.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "graphql.GetDrinksPoiPointOfInterest_ResultSetResultsPointOfInterestRdfs_commentLangString": {
            "type": "object",
            "properties": {
                "value": {
                    "description": "Literal value",
                    "type": "string"
                }
            }
        },
        "graphql.GetDrinksPoiPointOfInterest_ResultSetResultsPointOfInterestRdfs_labelLangString": {
            "type": "object",
            "properties": {
                "value": {
                    "description": "Literal value",
                    "type": "string"
                }
            }
        },
        "graphql.GetDrinksResponse": {
            "type": "object",
            "properties": {
                "poi": {
                    "$ref": "#/definitions/graphql.GetDrinksPoiPointOfInterest_ResultSet"
                }
            }
        },
        "graphql.GetEnjoyPoiPointOfInterest_ResultSet": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphql.GetEnjoyPoiPointOfInterest_ResultSetResultsPointOfInterest"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "graphql.GetEnjoyPoiPointOfInterest_ResultSetResultsPointOfInterest": {
            "type": "object",
            "properties": {
                "hasContact": {
                    "description": "L\u0026#039;agent à contacter pour affaires générales relatives à ce POI.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphql.GetEnjoyPoiPointOfInterest_ResultSetResultsPointOfInterestHasContactAgent"
                    }
                },
                "hasDescription": {
                    "description": "Description textuelle courte ou longue du POI pouvant être associée à une\naudience. Par exemple, un POI peut avoir une description dédiée aux écoles et\nune autre dédiée au grand public. Si une description n\u0026#039;a aucune audience\nrenseignée, on suppose qu\u0026#039;elle est dédiée à tous les publics.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphql.GetEnjoyPoiPointOfInterest_ResultSetResultsPointOfInterestHasDescription"
                    }
                },
                "lastUpdateDatatourisme": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rdfs_comment": {
                    "description": "Description de la ressource.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphql.GetEnjoyPoiPointOfInterest_ResultSetResultsPointOfInterestRdfs_commentLangString"
                    }
                },
                "rdfs_label": {
                    "description": "Etiquette courte décrivant la ressource.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphql.GetEnjoyPoiPointOfInterest_ResultSetResultsPointOfInterestRdfs_labelLangString"
                    }
                },
                "reducedMobilityAccess": {
                    "description": "Vrai si le produit propose un accès aux personnes à mobilité réduite",
                    "type": "array",
                    "items": {
                        "type": "boolean"
                    }
                }
            }
        },
        "graphql.GetEnjoyPoiPointOfInterest_ResultSetResultsPointOfInterestHasContactAgent": {
            "type": "object",
            "properties": {
                "foaf_homepage": {
                    "description": "L\u0026#039;adresse du site internet d\u0026#039;un Agent.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "schema_email": {
                    "description": "Un courriel, courrier électronique.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "schema_telephone": {
                    "description": "Un numéro de téléphone.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "graphql.GetEnjoyPoiPointOfInterest_ResultSetResultsPointOfInterestHasDescription": {
            "type": "object",
            "properties": {
                "dc_description": {
                    "description": "Description longue de la ressource.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphql.GetEnjoyPoiPointOfInterest_ResultSetResultsPointOfInterestHasDescriptionDc_descriptionLangString"
                    }
                }
            }
        },
        "graphql.GetEnjoyPoiPointOfInterest_ResultSetResultsPointOfInterestHasDescriptionDc_descriptionLangString": {
            "type": "object",
            "properties": {
                "value": {
                    "description": "Literal value",
                    "type": "string"
                }
            }
        },
        "graphql.GetEnjoyPoiPointOfInterest_ResultSetResultsPointOfInterestRdfs_commentLangString": {
            "type": "object",
            "properties": {
                "value": {
                    "description": "Literal value",
                    "type": "string"
                }
            }
        },
        "graphql.GetEnjoyPoiPointOfInterest_ResultSetResultsPointOfInterestRdfs_labelLangString": {
            "type": "object",
            "properties": {
                "value": {
                    "description": "Literal value",
                    "type": "string"
                }
            }
        },
        "graphql.GetEnjoyResponse": {
            "type": "object",
            "properties": {
                "poi": {
                    "$ref": "#/definitions/graphql.GetEnjoyPoiPointOfInterest_ResultSet"
                }
            }
        },
        "graphql.GetSleepPoiPointOfInterest_ResultSet": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphql.GetSleepPoiPointOfInterest_ResultSetResultsPointOfInterest"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "graphql.GetSleepPoiPointOfInterest_ResultSetResultsPointOfInterest": {
            "type": "object",
            "properties": {
                "hasContact": {
                    "description": "L\u0026#039;agent à contacter pour affaires générales relatives à ce POI.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphql.GetSleepPoiPointOfInterest_ResultSetResultsPointOfInterestHasContactAgent"
                    }
                },
                "hasDescription": {
                    "description": "Description textuelle courte ou longue du POI pouvant être associée à une\naudience. Par exemple, un POI peut avoir une description dédiée aux écoles et\nune autre dédiée au grand public. Si une description n\u0026#039;a aucune audience\nrenseignée, on suppose qu\u0026#039;elle est dédiée à tous les publics.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphql.GetSleepPoiPointOfInterest_ResultSetResultsPointOfInterestHasDescription"
                    }
                },
                "lastUpdateDatatourisme": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rdfs_comment": {
                    "description": "Description de la ressource.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphql.GetSleepPoiPointOfInterest_ResultSetResultsPointOfInterestRdfs_commentLangString"
                    }
                },
                "rdfs_label": {
                    "description": "Etiquette courte décrivant la ressource.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphql.GetSleepPoiPointOfInterest_ResultSetResultsPointOfInterestRdfs_labelLangString"
                    }
                },
                "reducedMobilityAccess": {
                    "description": "Vrai si le produit propose un accès aux personnes à mobilité réduite",
                    "type": "array",
                    "items": {
                        "type": "boolean"
                    }
                }
            }
        },
        "graphql.GetSleepPoiPointOfInterest_ResultSetResultsPointOfInterestHasContactAgent": {
            "type": "object",
            "properties": {
                "foaf_homepage": {
                    "description": "L\u0026#039;adresse du site internet d\u0026#039;un Agent.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "schema_email": {
                    "description": "Un courriel, courrier électronique.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "schema_telephone": {
                    "description": "Un numéro de téléphone.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "graphql.GetSleepPoiPointOfInterest_ResultSetResultsPointOfInterestHasDescription": {
            "type": "object",
            "properties": {
                "dc_description": {
                    "description": "Description longue de la ressource.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphql.GetSleepPoiPointOfInterest_ResultSetResultsPointOfInterestHasDescriptionDc_descriptionLangString"
                    }
                }
            }
        },
        "graphql.GetSleepPoiPointOfInterest_ResultSetResultsPointOfInterestHasDescriptionDc_descriptionLangString": {
            "type": "object",
            "properties": {
                "value": {
                    "description": "Literal value",
                    "type": "string"
                }
            }
        },
        "graphql.GetSleepPoiPointOfInterest_ResultSetResultsPointOfInterestRdfs_commentLangString": {
            "type": "object",
            "properties": {
                "value": {
                    "description": "Literal value",
                    "type": "string"
                }
            }
        },
        "graphql.GetSleepPoiPointOfInterest_ResultSetResultsPointOfInterestRdfs_labelLangString": {
            "type": "object",
            "properties": {
                "value": {
                    "description": "Literal value",
                    "type": "string"
                }
            }
        },
        "graphql.GetSleepResponse": {
            "type": "object",
            "properties": {
                "poi": {
                    "$ref": "#/definitions/graphql.GetSleepPoiPointOfInterest_ResultSet"
                }
            }
        },
        "graphql.GetTotalPoiPointOfInterest_ResultSet": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                }
            }
        },
        "graphql.GetTotalResponse": {
            "type": "object",
            "properties": {
                "poi": {
                    "$ref": "#/definitions/graphql.GetTotalPoiPointOfInterest_ResultSet"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}